Made download button into re-usable component
