+++ src/components/download-button/downloadButton.module.css
@@ -1,0 +1,14 @@
+.downloadButton {
+  padding: 16px;
+  background-color: black;
+  color: white;
+  border-radius: 16px;
+  border-color: white;
+  font-size: large;
+  font-weight: bold;
+}
+
+.downloadButton:hover {
+  color: black;
+  background-color: white;
+}
